+++ src/main/java/groovejames/service/ProxySettings.java
@@ -1,12 +1,11 @@
 package groovejames.service;
 
-public class ProxySettings {
+import java.util.Objects;
 
-    private String host;
-    private int port;
+public class ProxySettings implements Comparable<ProxySettings> {
 
-    public ProxySettings() {
-    }
+    private final String host;
+    private final int port;
 
     public ProxySettings(String host, int port) {
         if (host == null)
@@ -19,40 +18,31 @@ public class ProxySettings {
         return host;
     }
 
-    public void setHost(String host) {
-        this.host = host;
-    }
-
     public int getPort() {
         return port;
-    }
-
-    public void setPort(int port) {
-        this.port = port;
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (o == null) return false;
+        if (!(o instanceof ProxySettings)) return false;
         ProxySettings that = (ProxySettings) o;
-        return port == that.port && !(host != null ? !host.equals(that.host) : that.host != null);
+        return port == that.port && Objects.equals(host, that.host);
     }
 
     @Override
     public int hashCode() {
-        int result = host != null ? host.hashCode() : 0;
-        result = 31 * result + port;
-        return result;
+        return Objects.hash(host, port);
     }
 
     @Override
     public String toString() {
-        final StringBuilder sb = new StringBuilder();
-        sb.append("ProxySettings");
-        sb.append("{host='").append(host).append('\'');
-        sb.append(", port=").append(port);
-        sb.append('}');
-        return sb.toString();
+        return String.format("%s:%d", host, port);
+    }
+
+    @Override
+    public int compareTo(ProxySettings other) {
+        return this.toString().compareTo(other.toString());
     }
 }